parse: remove overlap between output and error
The previous return signature of parse was
  (opts map[string]any, output string, err error)

where the meaning of output was clearly overlapping with the error,
causing complexity, difficult to understand and buggy code.

Now we return simply
  (opts map[string]any, err error)

where the error contains also what was previously in output.
This already made docopt to print better error messages
(see the fix in examples/subcommands/testdata/cli.failure.txt)
but we are not done yet.